// src/app/features/expenses/payment-type/models/payment-method-dto.model.ts
export interface PaymentMethodDto {
    id: number;
    name: string;
    description: string;
    isActive: boolean;
    isDeleted: boolean;
    createdBy: number | null;
    updatedBy: number | null;
    deletedBy: number | null;
    createdAt: Date | null;
    updatedAt: Date | null;
    deletedAt: Date | null;
  }